requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.16.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.2
